backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
openai==1.51.0
langgraph==0.2.28
langchain-core==0.3.12
langchain-openai==0.2.2
alembic==1.13.3
